Enable one build for arm, x86, and simd
