+++ parsearm.go
@@ -2,74 +2,28 @@ package main
 
 import (
 	"bufio"
-	"bytes"
 	"fmt"
 	"os"
-	"regexp"
 	"strconv"
 	"strings"
-	"unicode"
 )
 
-type Param struct {
-	Name  string
-	Type  string
-	Const bool
-}
-
-type Params []Param
-
-type Intrinsic struct {
-	OrgName     string
-	Name        string
-	AsmName     string
-	Instruction string
-	Description string
-	RetType     string
-	Params      Params
-	CpuID       string
-	Operation   string
-	Performance Perf
-	Package     string
-	cParam      *Param
-}
-
-type Timing struct {
-	Arch       string
-	Latency    float64
-	Throughput float64
-}
-
-type Perf map[string]Timing
-
-//var m64, m128, m256, m512, gen *os.File
-//var m64a, m128a, m256a, m512a, gena *os.File
-type pack struct {
-	goFile  *os.File
-	asmFile *os.File
-}
-
-const genimport = `import . "github.com/klauspost/intrinsics/arm"`
-
-func main() {
+func parsearm() {
+	genimport = `import . "github.com/klauspost/intrinsics/arm"`
 	f, err := os.Open("arm_neon-processed.h")
 	if err != nil {
 		panic(err)
 	}
 	defer f.Close()
 
-	parseHeader(f)
+	parseHeaderArm(f)
 	for _, pk := range packages {
 		pk.goFile.Close()
 		//pk.asmFile.Close()
 	}
 }
 
-func NewIntrinsic() *Intrinsic {
-	return &Intrinsic{}
-}
-
-func parseHeader(r *os.File) {
+func parseHeaderArm(r *os.File) {
 	s := bufio.NewScanner(r)
 	s.Split(bufio.ScanWords)
 	for {
@@ -88,11 +42,11 @@ func parseHeader(r *os.File) {
 				end := strings.Contains(t, "{")
 				switch n {
 				case 0:
-					in.RetType = fixType(t)
+					in.RetType = fixTypeArm(t)
 					n++
 				case 1:
 					in.OrgName = t
-					in.Name = fixFuncName(in.OrgName)
+					in.Name = fixFuncNameArm(in.OrgName)
 					fmt.Print(in.Name)
 					n++
 				case 2:
@@ -102,7 +56,7 @@ func parseHeader(r *os.File) {
 						cParam.Const = true
 						continue
 					}
-					cParam.Type = fixType(t)
+					cParam.Type = fixTypeArm(t)
 					n++
 				case 3:
 					if t == "const" {
@@ -133,25 +87,12 @@ func parseHeader(r *os.File) {
 					break
 				}
 			}
-			in.Finish()
+			in.FinishArm()
 		}
 	}
 }
 
-func (i Intrinsic) getPackage() string {
-	pk := strings.ToLower(i.CpuID)
-	if pk == "" {
-		pk = "misc"
-	}
-	if pk == "sse4.1" || pk == "sse4.2" {
-		pk = "sse4"
-	}
-	return pk
-}
-
-var packages = make(map[string]pack)
-
-func (i Intrinsic) getFiles() pack {
+func (i Intrinsic) getFilesArm() pack {
 	pk := i.getPackage()
 	p, ok := packages[pk]
 	if !ok {
@@ -181,7 +122,7 @@ func (i Intrinsic) getFiles() pack {
 	return p
 }
 
-func fixFuncName(in string) string {
+func fixFuncNameArm(in string) string {
 	out := CamelCase(in)
 	if strings.HasPrefix(in, "v") {
 		out = strings.TrimPrefix(out, "v")
@@ -190,7 +131,7 @@ func fixFuncName(in string) string {
 	return out
 }
 
-func fixType(r string) string {
+func fixTypeArm(r string) string {
 	pointer := false
 	if strings.Contains(r, "*") {
 		pointer = true
@@ -340,94 +281,7 @@ func fixType(r string) string {
 
 }
 
-var camelingRegex = regexp.MustCompile("[0-9A-Za-z]+")
-
-func CamelCase(src string) string {
-	byteSrc := []byte(src)
-	chunks := camelingRegex.FindAll(byteSrc, -1)
-	for idx, val := range chunks {
-		if idx > 0 {
-			chunks[idx] = bytes.Title(val)
-		}
-	}
-	return string(bytes.Join(chunks, nil))
-}
-
-func (p Params) HasParam(s string) bool {
-	for _, param := range p {
-		if param.Name == s {
-			return true
-		}
-	}
-	return false
-}
-
-var usedNames = make(map[string]struct{})
-
-func (in *Intrinsic) fixup() {
-	in.Package = in.getPackage()
-	name := in.Name
-	next := 2
-	for {
-		_, ok := usedNames[in.Package+name]
-		if !ok {
-			break
-		}
-		fmt.Println("Warning: Name conflict:", in.Package+"/"+name+"(...)")
-		name = fmt.Sprintf("%s%d", in.Name, next)
-		next++
-
-	}
-	usedNames[in.Package+name] = struct{}{}
-	in.Name = name
-	a := []rune(name)
-	a[0] = unicode.ToLower(a[0])
-
-	in.AsmName = string(a)
-	for i, param := range in.Params {
-		switch param.Name {
-		case "type":
-			in.Params[i].Name = "typ"
-		case "func":
-			in.Params[i].Name = "fnc"
-		case "imm8":
-			in.Params[i].Type = "byte"
-		}
-		if param.Type == "" {
-			in.Params[i].Type = "uintptr"
-		}
-	}
-
-}
-
-func (in Intrinsic) shouldSkip() bool {
-	if in.RetType == "uintptr" {
-		return true
-	}
-	if strings.HasPrefix(in.RetType, "*") {
-		return true
-	}
-	for _, param := range in.Params {
-		if param.Type == "uintptr" {
-			return true
-		}
-	}
-	return false
-}
-
-func (in Intrinsic) shouldRework() bool {
-	if strings.HasPrefix(in.RetType, "*") {
-		return true
-	}
-	for _, param := range in.Params {
-		if strings.HasPrefix(param.Type, "*") {
-			return true
-		}
-	}
-	return false
-}
-
-func (in Intrinsic) hasImmediate() bool {
+func (in Intrinsic) hasImmediateArm() bool {
 	for _, param := range in.Params {
 		if param.Const {
 			return true
@@ -436,14 +290,14 @@ func (in Intrinsic) hasImmediate() bool {
 	return false
 }
 
-func (in Intrinsic) Finish() {
+func (in Intrinsic) FinishArm() {
 	if in.Name == "" {
 		return
 	}
 
 	in.fixup()
 
-	out := in.getFiles().goFile
+	out := in.getFilesArm().goFile
 
 	if in.shouldSkip() {
 		fmt.Fprintf(out, "\n// Skipped: %s. Contains pointer parameter.\n\n", in.OrgName)
@@ -467,7 +321,7 @@ func (in Intrinsic) Finish() {
 	if rework {
 		fmt.Fprintln(out, "//\n// FIXME: Will likely need to be reworked (has pointer parameter).")
 	}
-	if in.hasImmediate() {
+	if in.hasImmediateArm() {
 		fmt.Fprintln(out, "//\n// FIXME: Requires compiler support (has immediate)")
 	}
 	fmt.Fprint(out, "func ", in.Name, "(")
@@ -504,9 +358,9 @@ func (in Intrinsic) Finish() {
 	}
 	for i, param := range in.Params {
 		pn := param.Name
-		if param.getNative() != "" {
+		if param.getNativeArm() != "" {
 			// cast
-			pn = param.getNative() + "(" + pn + ")"
+			pn = param.getNativeArm() + "(" + pn + ")"
 		}
 		fmt.Fprint(out, pn)
 		if i != len(in.Params)-1 {
@@ -527,24 +381,24 @@ func (in Intrinsic) Finish() {
 	params = []string{}
 	for _, param := range in.Params {
 		typ := param.Type
-		ot := param.getNative()
+		ot := param.getNativeArm()
 		if ot != "" {
 			typ = ot
 		}
 		params = append(params, fmt.Sprint(param.Name, " ", typ))
 	}
-	if retparam.getNative() != "" {
-		retparam = Param{Type: retparam.getNative()}
+	if retparam.getNativeArm() != "" {
+		retparam = Param{Type: retparam.getNativeArm()}
 	}
 	fmt.Fprint(out, strings.Join(params, ", "), ") ", retparam.Type, "\n\n")
 
 	//emptyType(in.RetType)
 	// ASM
-	out = in.getFiles().asmFile
+	out = in.getFilesArm().asmFile
 	fmt.Fprint(out, "// func ", in.AsmName, "(")
 	fmt.Fprint(out, strings.Join(params, ", "), ") ", retparam.Type, "\n")
 	fmt.Fprint(out, "TEXT ·"+in.AsmName+"(SB),7,$0\n")
-	load, off := in.Params.getAsm()
+	load, off := in.Params.getAsmArm()
 	fmt.Fprintln(out, load)
 	retparam = Param{Type: in.RetType}
 
@@ -580,12 +434,12 @@ func (in Intrinsic) Finish() {
 			}
 		}
 
-		if retparam.getSize() > 0 && retparam.getReg(0) == "R8" {
+		if retparam.getSizeArm() > 0 && retparam.getReg(0) == "R8" {
 			fmt.Fprint(out, "\tMOV"+retparam.getPostfix()+" $0, ret+"+strconv.Itoa(off)+"(FP)\n")
-		} else if retparam.getSize() > 8 {
+		} else if retparam.getSizeArm() > 8 {
 			fmt.Fprint(out, "\tMOV"+retparam.getPostfix()+" "+rreg+", ret+"+strconv.Itoa(off)+"(FP)\n")
 		} else {
-			fmt.Fprintf(out, "\t// Return size: %d\n", retparam.getSize())
+			fmt.Fprintf(out, "\t// Return size: %d\n", retparam.getSizeArm())
 		}
 
 	}
@@ -593,7 +447,7 @@ func (in Intrinsic) Finish() {
 	fmt.Fprint(out, "\n")
 }
 
-func (p Param) getSize() int {
+func (p Param) getSizeArm() int {
 	t := p.Type
 	tsplit := strings.Split(t, ".")
 	if len(tsplit) > 1 {
@@ -621,7 +475,7 @@ func (p Param) getSize() int {
 	return 0
 }
 
-func (p Param) getNative() string {
+func (p Param) getNativeArm() string {
 	t := p.Type
 	tsplit := strings.Split(t, ".")
 	if len(tsplit) > 1 {
@@ -692,59 +546,11 @@ func (p Param) getNative() string {
 	return ""
 }
 
-func (p Param) getPostfix() string {
-	size := p.getSize()
-	switch size {
-	case 1:
-		return "B"
-	case 2:
-		return "W"
-	case 4:
-		return "L"
-	case 8:
-		return "Q"
-	case 16:
-		return "OU"
-	}
-	return ""
-}
-
-func (p Param) getReg(n int) string {
-	if n > 7 {
-		return ""
-	}
-	t := p.Type
-	tsplit := strings.Split(t, ".")
-	if len(tsplit) > 1 {
-		t = tsplit[1]
-	}
-
-	switch t {
-	case "M64", "M64i":
-		return "M" + strconv.Itoa(n)
-	case "M128", "M128i", "M128d":
-		return "X" + strconv.Itoa(n)
-	case "M256", "M256i", "M256d":
-		return "Y" + strconv.Itoa(n)
-	case "M512", "M512i", "M512d":
-		return "Z" + strconv.Itoa(n)
-	case "int", "uint", "int64", "uint64", "float64", "uintptr", "Mmask64":
-		return "R" + strconv.Itoa(8+n)
-	case "int32", "uint32", "float32", "Mmask32":
-		return "R" + strconv.Itoa(8+n)
-	case "int16", "uint16", "Mmask16":
-		return "R" + strconv.Itoa(8+n)
-	case "int8", "uint8", "Mmask8", "byte":
-		return "R" + strconv.Itoa(8+n)
-	}
-	return ""
-}
-
-func (p Params) getAsm() (string, int) {
+func (p Params) getAsmArm() (string, int) {
 	out := ""
 	atbytes := 0
 	for i, param := range p {
-		size := param.getSize()
+		size := param.getSizeArm()
 		if size == 0 {
 			return "\t// FIXME: Unimplemented. Unknown size of type " + param.Type + "\n", 0
 		}
@@ -767,102 +573,4 @@ func (p Params) getAsm() (string, int) {
 		atbytes += size
 	}
 	return out, atbytes
-}
-
-func emptyType(s string) string {
-	if strings.HasPrefix(s, "int") {
-		return "0"
-	}
-	if strings.HasPrefix(s, "uint") {
-		return "0"
-	}
-	if strings.HasPrefix(s, "Mmask") {
-		return s + "(0)"
-	}
-	if strings.HasPrefix(s, "float") {
-		return "0.0"
-	}
-	switch s {
-	case "string":
-		return `""`
-	case "byte":
-		return "0"
-	case "unsafe.Pointer":
-		return s + "(uintptr(0))"
-	default:
-		return s + "{}"
-	}
-}
-
-// WrapString wraps the given string within lim width in characters.
-//
-// Wrapping is currently naive and only happens at white-space. A future
-// version of the library will implement smarter wrapping. This means that
-// pathological cases can dramatically reach past the limit, such as a very
-// long word.
-func WrapString(s string, lim uint) string {
-	// Initialize a buffer with a slightly larger size to account for breaks
-	init := make([]byte, 0, len(s))
-	buf := bytes.NewBuffer(init)
-
-	var current uint
-	var wordBuf, spaceBuf bytes.Buffer
-	var skipline bool
-
-	for _, char := range s {
-		if char == '\n' {
-			if wordBuf.Len() == 0 {
-				if current+uint(spaceBuf.Len()) > lim {
-					current = 0
-				} else {
-					current += uint(spaceBuf.Len())
-					spaceBuf.WriteTo(buf)
-				}
-				spaceBuf.Reset()
-			} else {
-				current += uint(spaceBuf.Len() + wordBuf.Len())
-				spaceBuf.WriteTo(buf)
-				spaceBuf.Reset()
-				wordBuf.WriteTo(buf)
-				wordBuf.Reset()
-			}
-			buf.WriteRune(char)
-			current = 0
-			skipline = false
-		} else if unicode.IsSpace(char) {
-			if current == 0 && spaceBuf.Len() > 2 {
-				skipline = true
-			}
-
-			if spaceBuf.Len() == 0 || wordBuf.Len() > 0 {
-				current += uint(spaceBuf.Len() + wordBuf.Len())
-				spaceBuf.WriteTo(buf)
-				spaceBuf.Reset()
-				wordBuf.WriteTo(buf)
-				wordBuf.Reset()
-			}
-
-			spaceBuf.WriteRune(char)
-		} else {
-
-			wordBuf.WriteRune(char)
-
-			if current+uint(spaceBuf.Len()+wordBuf.Len()) > lim && uint(wordBuf.Len()) < lim && !skipline {
-				buf.WriteRune('\n')
-				current = 0
-				spaceBuf.Reset()
-			}
-		}
-	}
-
-	if wordBuf.Len() == 0 {
-		if current+uint(spaceBuf.Len()) <= lim {
-			spaceBuf.WriteTo(buf)
-		}
-	} else {
-		spaceBuf.WriteTo(buf)
-		wordBuf.WriteTo(buf)
-	}
-
-	return buf.String()
 }
